[skip ci] image_aggregator: Match std::min arg types am: efbfbc79f3 am: efc0817107 am: cfcacac64c am: a8831d162c
Original change: https://android-review.googlesource.com/c/device/google/cuttlefish/+/1736194

GitOrigin-RevId: 79f96a3874a5e00828e944c1fa2b6795feb136e1

+++ staging/host/libs/image_aggregator/image_aggregator.cc
@@ -168,7 +168,7 @@ std::uint64_t ExpandedStorageSize(const std::string& file_path) {
 
   // Try to read the disk in a nicely-aligned block size unless the whole file
   // is smaller.
-  constexpr size_t MAGIC_BLOCK_SIZE = 4096;
+  constexpr uint64_t MAGIC_BLOCK_SIZE = 4096;
   std::string magic(std::min(file_size, MAGIC_BLOCK_SIZE), '\0');
   if (!android::base::ReadFully(fd, magic.data(), magic.size())) {
     PLOG(FATAL) << "Fail to read: " << file_path;
